srcs/backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
redis>=4.5
django-redis>=5.2
python-decouple>=3.8
channels>=4.0
channels-redis>=4.0
daphne>=4.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=9.0
